Refactor `text_controller` logic
